+++ sandbox/grommet-app/src/pages/layouts/kinds/OpsRamp/ResourceDetails.tsx
@@ -4,9 +4,9 @@ import {
   Button,
   Menu,
   Text,
-  ResponsiveContext,
   Layer,
   Heading,
+  ResponsiveContext,
 } from 'grommet';
 import { LinkNext, StatusGoodSmall } from 'grommet-icons';
 import ContentPane from '../../../../components/ContentPane';
@@ -91,10 +91,8 @@ export const ResourceDetails: React.FC<ResourceDetailsProps> = ({
   const metricSize = ['xsmall', 'small', 'medium'].includes(breakpoint)
     ? 'small'
     : 'medium';
-
   const content = (
     <ContentPane
-      gap="small"
       heading={
         <Box align="center" direction="row" gap="medium">
           <Button onClick={onClose} icon={<LinkNext />} />
@@ -107,19 +105,20 @@ export const ResourceDetails: React.FC<ResourceDetailsProps> = ({
       actions={<ResourceActions onClose={onClose} />}
       skeleton={undefined}
       animation={animation}
+      height="fit-content"
       {...rest}
     >
       <ResourceSections size={metricSize} />
     </ContentPane>
   );
 
-  return layer ? (
-    <Layer full modal={false}>
-      {content}
-    </Layer>
-  ) : (
-    content
-  );
-};
+  if (layer) {
+    return (
+      <Layer full modal={false} animation={animation}>
+        {content}
+      </Layer>
+    );
+  }
 
-export default ResourceDetails;
+  return content;
+};
