+++ sandbox/grommet-app/src/pages/layouts/kinds/OpsRamp/ServersTable.tsx
@@ -4,7 +4,6 @@ import {
   Box,
   Button,
   DataFilters,
-  Grid,
   Toolbar,
   DataSearch,
   Data,
@@ -22,10 +21,16 @@ import {
 } from 'grommet-icons';
 import opsRamp from '../../../../mockData/opsRamp.json';
 
-import { ResourceDetails } from './ResourceDetails';
 import { QuickFilters } from './QuickFilters';
 
-export const ServersTable: React.FC = () => {
+interface ServersTableProps {
+  showResultDetails: boolean;
+  setShowResultDetails: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export const ServersTable: React.FC<ServersTableProps> = ({
+  setShowResultDetails,
+}) => {
   const [total, setTotal] = useState(0);
   const [result, setResult] = useState([]);
   const [quickFilter, setQuickFilter] = useState('');
@@ -33,8 +38,6 @@ export const ServersTable: React.FC = () => {
   const [down, setDown] = useState(0);
   const [unknown, setUnknown] = useState(0);
   const [undefinedState, setUndefinedState] = useState(0);
-  const [showResultDetails, setShowResultDetails] = useState(false);
-  const breakpoint = useContext(ResponsiveContext);
 
   useEffect(() => {
     const servers = opsRamp.servers || [];
@@ -111,124 +114,62 @@ export const ServersTable: React.FC = () => {
     },
   ];
 
-  const gridColumns = ['xsmall', 'small'].includes(breakpoint)
-    ? ['flex', 'auto']
-    : ['auto', 'auto'];
-
-  const rows = ['auto'];
-  const columnsGap = {
-    xsmall: 'small',
-    small: 'small',
-    medium: 'medium',
-    large: 'large',
-    xlarge: 'large',
-  };
-  const rowsGap = {
-    xsmall: 'xsmall',
-    small: 'xsmall',
-    medium: 'small',
-    large: 'medium',
-    xlarge: 'medium',
-  };
-
-  const gap = {
-    row: rowsGap[breakpoint],
-    column: columnsGap[breakpoint],
-  };
-
-  const areasDefault = [
-    ['quickfilters', 'resourcedetails'],
-    ['datatable', 'resourcedetails'],
-  ];
-
-  const areasSmall = [
-    ['quickfilters', 'unassigned'],
-    ['datatable', 'unassigned'],
-  ];
-
-  const areas = ['xsmall', 'small'].includes(breakpoint)
-    ? areasSmall
-    : areasDefault;
-
   return (
-    <Grid
-      rows={rows}
-      columns={showResultDetails ? gridColumns : 'full'}
-      areas={showResultDetails ? areas : undefined}
-      gap={{ row: gap.row, column: gap.column }}
+    <Data
+      data={result}
+      total={total}
+      filteredTotal={result.length}
+      defaultView={{
+        search: '',
+        sort: { property: 'name', direction: 'asc' },
+        step: 10,
+      }}
+      properties={{
+        name: {
+          label: 'Name',
+        },
+        'ip address': {
+          label: 'IP Address',
+        },
+        model: {
+          label: 'Model',
+        },
+        state: {
+          label: 'State',
+        },
+      }}
     >
-      <Data
-        data={result}
-        total={total}
-        filteredTotal={result.length}
-        defaultView={{
-          search: '',
-          sort: { property: 'name', direction: 'asc' },
-          step: 10,
-        }}
-        properties={{
-          name: {
-            label: 'Name',
-          },
-          'ip address': {
-            label: 'IP Address',
-          },
-          model: {
-            label: 'Model',
-          },
-          state: {
-            label: 'State',
-          },
-        }}
+      <Box
+        round="small"
+        background="background-front"
+        pad={{ top: 'small', bottom: 'medium', horizontal: 'medium' }}
+        gap="medium"
+        flex={false}
       >
-        <Box
-          round="small"
-          background="background-front"
-          pad={{ top: 'small', bottom: 'medium', horizontal: 'medium' }}
-          gap="medium"
-        >
-          <Box gridArea="quickfilters">
-            <QuickFilters
-              value={quickFilter}
-              setValue={setQuickFilter}
-              counts={{ up, down, unknown, undefinedState }}
-            />
-          </Box>
-          <Box pad={{ vertical: 'small' }} overflow="auto" gridArea="datatable">
-            <Toolbar>
-              <DataSearch />
-              <DataFilters layer />
-              <Button kind="toolbar" icon={<FormUpload />} />
-              <Button kind="toolbar" icon={<SettingsOption />} />
-            </Toolbar>
-            <DataTable
-              aria-describedby="server-table"
-              onSelect={() => {}}
-              columns={columns}
-              sortable
-            />
-            <Pagination
-              summary
-              stepOptions
-              border="top"
-              pad={{ vertical: 'xsmall', left: 'small' }}
-            />
-          </Box>
-        </Box>
-      </Data>
-      <Box gridArea="resourcedetails">
-        {showResultDetails && (
-          <ResourceDetails
-            layer={['xsmall', 'small', 'medium'].includes(breakpoint)}
-            onClose={() => setShowResultDetails(false)}
-            animation={
-              ['xsmall', 'small', 'medium'].includes(breakpoint)
-                ? false
-                : ['slideLeft', 'fadeIn']
-            }
-          />
-        )}
+        <QuickFilters
+          value={quickFilter}
+          setValue={setQuickFilter}
+          counts={{ up, down, unknown, undefinedState }}
+        />
+        <Toolbar>
+          <DataSearch />
+          <DataFilters layer />
+          <Button kind="toolbar" icon={<FormUpload />} />
+          <Button kind="toolbar" icon={<SettingsOption />} />
+        </Toolbar>
+        <DataTable
+          aria-describedby="server-table"
+          onSelect={() => {}}
+          columns={columns}
+          sortable
+        />
+        <Pagination
+          summary
+          stepOptions
+          border="top"
+          pad={{ vertical: 'xsmall', left: 'small' }}
+        />
       </Box>
-    </Grid>
+    </Data>
   );
 };
